requirements.txt
snowflake-connector-python==3.7.0
pandas==2.2.0
requests==2.31.0
python-dotenv==1.0.0